damn

+++ 1352B.cpp
@@ -13,12 +13,10 @@ int main() {
     int t;
     cin >> t;
     while(t--){
-        long long n ;
+        int n ;
         int k;
         cin >> n >> k;
 
-        // all 1s 
-        // (k -1 ) 1s , sand then n - k +1 
         if( (n-k+1)>0 && (n-k+1)%2 == 1  ){
             cout << "YES" << '\n' ;
             for (int i = 0; i < k-1; i++) {
@@ -26,7 +24,6 @@ int main() {
             }
             cout << n-k+1 << '\n' ;
         }
-        // ( k- 1) 2's and then n - 2k +2
         else if( (n-2*k+2)>0 && (n-2*k+2)%2 == 0  ){
             cout << "YES" << '\n' ;
             for (int i = 0; i < k-1; i++) {
